Updated openstack/openstack Project: openstack-infra/project-config  bbf269dd1aafc4dc1870c593c00d7e7e046a4042
Add a 'base-minimal' job with it's set of playbooks

The 'base-minimal' job has for purpose to be what we consider to be
the smallest possible set of roles/tasks to be applied on nodes
regardless of the job context.

Inheriting from this minimal job provides the ability to run roles
or tasks that would have otherwise been included by default only on
a need basis.

Change-Id: I5c255a6f9d2adb91fbc222742bd2249f126ca982