Updated openstack/openstack Project: openstack/neutron  e711efc7db005ff9ee903bd24f7a266053a4f90f
fullstack: skip test_mtu_update when DHCP agent is not in rootns

The test case requires access to devices created by DHCP agent. When the
agent is deployed with linuxbridge agent, we start it in a host-
namespace, and then it creates its own qdhcp- namespaces for dhcp ports
inside the host- namespace.

This could be solved by calling:
ip netns exec host-* ip netns exec qdhcp-* ip link show ...
but ip_lib doesn't support constructing such commands.

Of course, we could extend ip_lib to support this use case, but
currently we have larger dragons to fight. We will revisit the skip once
all agents are moved to namespaces as planned by jlibosva@.

This patch merely disables the test case if DHCP agent is running in a
namespace. We may revisit it when and if we feel comfortable making
significant changes to ip_lib.

Change-Id: Ieea83bf2959654f98b1dc4d7aa17138d0626ec40
Related-Bug: #1671634