Updated openstack/openstack Project: openstack-infra/project-config  b8e72b49a0ba9bebc90780e3af9a3d1512c46891
Add the emit-ara-html to the base-minimal role

The intent was to test the emit-ara-html role as part of the ongoing
work for testing the zuul base playbooks but it turns out it's a bit
more complicated as it's intended to run on the executor.

Let's add ara in for the time being and perhaps figure out another
way to test it later.

Change-Id: I1c6da62585b93ca28460eabed4eebe2d020ba82b