Updated openstack/openstack Project: openstack/cinder  fe7e123e781ec87f60b516c475dd9c8591fd0f7d
Implement keymgr list() method

Castellan KeyManager has added an abstract list() method starting
in release 0.13.0. We are currently safe since upper constraints
cap it at 0.12.0, but we will want to raise those constraints
soon.

This method isn't actually used in the Cinder code, but since it's
an ABC abstract method we must implement it in our ConfKeyMgr. This
just adds a basic skeleton so we don't blow up on missing the method.

Change-Id: Ie2b2aef40b4de36d011c7efa0ab5af5c9581a63d
Closes-bug: #1715451