Updated openstack/openstack Project: openstack/nova  1afc9c5a728c414e8017fd55dddad24de6c84a56
Ensure instance mapping is updated in case of quota recheck fails

If an instance fails to successfully pass the quota recheck, it will
raise a TooManyInstances exception, however, it will not hit the
code which saves the instance mapping, leaving an instance with no
assigned cell in the mapping table and no BuildRequest as it is
removed by _cleanup_build_artifacts.

This patch adds a test to make sure that an instance has the correct
cell mapping if it fails in the quota recheck phase.  In addition, it
uses the cell_mapping_cache dictionary to set the correct cell
mapping before marking the instance as ERROR.

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>
Co-Authored-By: melanie witt <melwittt@gmail.com>
Closes-Bug: #1715462

Change-Id: I7ecb5feb47a5f358cd51bde87b75a3a6141b5b12