Updated openstack/openstack Project: openstack/cinder  f2319a0ff34bc084261e8a71461f00776080a68a
Fix attachments on volume show when volume is attaching

With the new v3 cinder attach flow, nova calls attachment_update
and then attachment_complete. During that time the volume
has 'attaching' status. Because of that status, cinder show volume
will show all the attachments as empty, even if the volume has
attachments. Currently this will only happen during migration,
but when multi-attach is working, this will be a general issue.

The change is to the show volume flow. It will no longer look at the
volume attach status, but only at the individual attachment status's.
Any attachment that has status of attached will be returned by show.

(This will only be an issue in Queens when the new multi-attach flow
is used. There is no need for this change in Pike.)

Partially Implements: blueprint cinder-new-attach-apis
Change-Id: I8a2bff5a668ec58ee80c192cb72326f2b3599c39
Closes-Bug: 1713521