Updated openstack/openstack Project: openstack-infra/system-config  cb51d5b33212a05948433d37da0e411d0f8e1de6
Force OVH to use ipv4 in shade

OVH instances are getting ipv6 addresses now. This is great except that
the only info on these addresses seems to be in the nova/neutron api and
not via ipv6 router advertisements or config drive metadata. Until we
sort out a programmatic way to configure ipv6 on the instances at boot
lets not attempt to use ipv6 when connecting to OVH instances.

Change-Id: I48e773b9f29fae9a15f6c95fc64831ab0749cc60