Updated openstack/openstack Project: openstack/openstack-manuals  929da12725e7b3ba12092d5cf847340f7f97b122
do not require trailing slash in all redirects

Update the redirect rules to allow them to work if the URL does not
have a trailing slash. In some cases we need to use separate rules
because some of our project names are prefixes for other project
names, so the rules to redirect glance.* would pick up glance_store
and redirect to the wrong place.

Change-Id: I69290b9239acab44dc5e523a4feef4a2b48ed304
Signed-off-by: Doug Hellmann <doug@doughellmann.com>