Updated openstack/openstack Project: openstack-infra/project-config  f13a7985264340ed503dc8c8b5982218ee8efbd8
Add wheel mirror jobs

We need to run periodic jobs to build the wheel mirrors.

This removes them from the v2 config because we really should not run
these jobs from both systems at the same time and once this lands it'll
be live.

It copies the scripts from project-config to the homedir, so we can
ditch the puppetted copies of them.

Change-Id: I87a78509a97d53eb627a84c646a6a3f0a227d9a1