Updated openstack/openstack Project: openstack-infra/project-config  e072dcafeef0b85872e7c55ab218966cf87e9989
Treat wheel builders more like build nodes

Pass the wheel keytab across and aklog with it.

Also, split the build role into a build and a copy so that we only have
the token when we need it, and not while we're running the wheel builds
themselves. And split release from build so that we can release the
mirror once for a consistent state across python2 and python3.

Change-Id: Id30a99f5b515dd520d013fabbd356e32c82030a0