Updated openstack/openstack Project: openstack-infra/project-config  43d8cfdf6030baabd5268a76a6b850e2b5a45310
Add all the rest of the pipelines

These are all of the rest of the v2 pipelines.

Also, removed reverify from check and added the require-approval stanza
from v2's check.

The experimental pipelines are included even though we don't actually
need them in v3 with job-dependencies, to ease migration.

The silent pipeline was omitted, as it's not currently used.

The email settings were included and uncommented, as we should start
figuring out if email is broken.

Change-Id: I27fa870af279f91852697c030497fad55c0cd3e0