Updated openstack/openstack Project: openstack/heat  0de1e372c70e2b3a4a78207109a25cc9801c80ec
Minor doc fix: tmpl_diff argument for handle_update

The description for tmpl_diff argument in handle_update function
was incomplete and could be misconstrued as referring to the
values in the old definition. This documentation fix rectifies
that issue. Also, corrected the name "templ_diff" to "tmpl_diff".

Related-Bug: 1709147
Change-Id: Iaa55aeb14cf836d0dc589bdbe711fa29f6ef6080