Updated openstack/openstack Project: openstack-infra/project-config  090956404bed01fd92505514cb91d27266d45265
Move gate-osc-placement-dsvm-functional-ubuntu-xenial to check

This moves the gate-osc-placement-dsvm-functional-ubuntu-xenial
job out of the experimental queue and makes it a non-voting job
in the check queue for the osc-placement repo.

This will be used to test changes proposed to the osc-placement
plugin itself using a functional devstack setup.

Related blueprint bp/placement-osc-plugin

Change-Id: Iba4bf348dc223e2ab04b5e5067d4fb4cd5e4cea9