Updated openstack/openstack Project: openstack/nova  def81cace8fbdfafb8488d0e400c99935164481f
Move libvirt usages of chown to privsep.

A nice simple example of how to move things to privsep (and the new set
of helpers). In a few of these cases I think a more complicated re-write
is actually required, but I've put TODOs there are will do those in a
followup patch.

Change-Id: Ibb6ef001e3f2add459b0e37dfbd9f51c9eff2eb7