Updated openstack/openstack Project: openstack-infra/project-config  39716d10cc7b1cb6d4435da77779483a3891d7aa
Consume validate_host parameters from site-variables

The settings are in the site-variables.yaml file, we don't need to
provide them in the job anymore.

Also, move validate-host earlier so that we collect the generated
inventory file and check network connectivity before we attempt to push
git repos over the wire.

Depends-On: I632a32fdfac4bfe57eb269ac8e183fb8df34d48f
Change-Id: I1131eb47f836f8c86b7e9e6329199612243873aa