Updated openstack/openstack Project: openstack/neutron  75d98fcd79a716e2e795ad9b583209eefc009f25
Fix generation of thousands of DHCP tap interfaces

As reported in the bug, there may be an case where an empty
namespace file in /run/netns, but the namespace not
actually exist. In such case the DHCP agent throws an error
when pluging the interface in the dhcp namespace.
This may also result in many tap interfaces
getting generated in OVS bridge or Linux bridge.

This patch fixes the above bug by unpluging the tap device
in the bridge if exception occurs, this can prevents the tap
interfaces generate.

Co-Authored-By: Brian Haley <bhaley@redhat.com>

Change-Id: I4a197edd180887ad36317ddb2f0c0e7bd2e34e30
Closes-Bug: #1561695