Updated openstack/openstack Project: openstack-infra/shade  449bc35eddba89a9d75fef300f0b4ba25fbd9ad9
Switch to using stestr

stestr is a fork of testrepository, designed specifically to concentrate
on being a dedicated test runner. [1] The testrepository project is
basically not active anymore and has several long standing bugs and
useablility issues. stestr is still actively maintained and fixes a
large number of those issues with testrepository.

[1] https://github.com/mtreinish/stestr

Change-Id: I9b8049fd2c78d8f90aed5a8b35c0a9b40ad8c463