Updated openstack/openstack Project: openstack-infra/project-config  4ef808962f0f04061bc8369e4e3d64848b9223ba
be more narrow about doc skips in nova

Nova is skipping functional tests for docs changes, which mostly is
good, however some of the files in the docs trees are api samples,
which are actually used in tests beyond docs. A breaking change to api
samples landed recently that we missed because of this.

We can narrow to just doc/source/ exclusion and life should be good.

Change-Id: I36892cb5bdee52f4dfbb6a5764c03ceea78b9dc4