Updated openstack/openstack Project: openstack/tripleo-heat-templates  e0867ddffae5fb7d0ebe4f536f776fac86682b69
Stricter heat template version validation

The current check just validates that the heat template versions are using the
release alias instead of the date format.

This patch add stricter validation to ensure the heat template versions used
are supported by the release.

Also emits a warning if the most current template
version is not used.

Change-Id: I9e1d9b05fd5aa91c6a26d032043f386b6a9b072d