Updated openstack/openstack Project: openstack-infra/nodepool  881b2fcba6be3654c3ca7aa3eaced110c2414e3a
Add SSH Host Key Verifier Strategy

Since version 1.15, the Jenkins SSH Slave Plugin has added a requirement
for a SSH Host Key Verification Strategy. This defaults to requiring
that the host key is already accepted by the Jenkins SSH environment.

This path reactivates the old and expected behavior of accepting, but
throwing a warning in the job logs whenever a host is SSHed into that
has an unknown SSH host key.

Change-Id: I2316cac0b2bb6ff2f0f4abc7dfd5307d80edf1b3