Updated openstack/openstack Project: openstack/neutron  1323b6cdbd2f01a32a5ee518e9ecf5abe2b24578
DVR: Multiple csnat ports created when RouterPort table update fails

When router interfaces are added to DVR router, if the router has
gateway configured, then the internal csnat ports are created for
the corresponding router interfaces.
We have seen recently after the csnat port is created if the
RouterPort table update fails, there is a DB retry that is happening
and that retry operation is creating an additional csnat port.
This additional port is not getting removed automatically when the
router interfaces are deleted.
This issue is seen when testing with a simple heat template as
per the bug report.

This patch fixes the issue by calling the RouterPort create with
delete_port_on_error context.

Change-Id: I916011f2200f02556ebb30bce30e349a8023602c
Closes-Bug: #1709774