Updated openstack/openstack Project: openstack/requirements  9e771e7ba73a35b58d80a0c53020b10f7d4b4ad6
Update lower bounds for oslo.*, os-* and *client to recent versions

Most of them were neglected and pointed to versions released
a year ago and more, which are typically not installable anymore
(they used to have a pbr<2.0 requirement which is incompatible
with the pbr>=2.0 requirement we enforce globally).

So for all of the openstack affine projects, update to a release
that is less than 4 months old (in some cases even newer versions
were required upon trial and error).

Change-Id: I69472a6a99b857e9e5dad9052eda7172ae2c9525