Updated openstack/openstack Project: openstack/horizon  4277b078b9bf2556b0fcdbd355ae465444a1e75f
Make @memoize thread-aware

If 2 parallel threads simultaneously make an expensive call to a
memoized function with arguments for which the returned value wasn's
cached before, 2 calls are effectively made in parallel.

It's better to avoid such a situation and if one thread already
initiated an expensive call, block on it in another thread and once
the call in the first thread is completed, unblock and reuse the value
which is cached now.

Co-Authored-By: Vladislav Kuzmin <vkuzmin@mirantis.com>
Change-Id: Id216e7b9c1bd836ee70a36ecb378e7d51d3821d2