Updated openstack/openstack Project: openstack/barbican  f533c0bda4cc1fb4345552c6827628c813dd5f76
allow redirects in .htaccess files on the static web servers

As part of allowing projects to establish redirects when they move
documentation pages around, we need to let them use .htaccess files.
We only want to allow Redirect and RedirectMatch directives,
though. This change restricts the directives on docs.o.o and adds them
on the static servers used for drafts published during a build (to
allow testing of the redirects).

For more details see:
 http://lists.openstack.org/pipermail/openstack-dev/2017-July/120240.html

Change-Id: I754c8016ca2a475e74023017530264e7ae331557