Updated openstack/openstack Project: openstack-infra/storyboard  1cfb51aa65a26a6129f8ee92896cebc8b6e88ffb
Pin os-testr to version 0.8.2

This commit pins the os-testr version to 0.8.2, as more recent changes
to the backend mean that some storyboard modules are not found. In the
future, we should adjust storyboard to work with the more recent
os-testr version, either by changing paths to those modules or
by explicitly pointing to them.

Change-Id: Iad68613d294179548d4ca73d75df699b05e7d0d7