Updated openstack/openstack Project: openstack-infra/storyboard  42511bf954203776a52a2007c114c821b615d3dc
Minimise database queries when resolving cards

This commit modifies the code which resolves the contents of cards
in boards and worklists such that instead of making a database
query per-card, two queries are made to obtain all tasks and stories
that there are cards for in the board/worklist.

The result of these queries is passed into the resolution function
in order to allow fast lookup of the underlying story/task for a card.

This will lead to much faster loading times for boards containing
large numbers of cards.

Change-Id: I189a1d9823a7d251af641cf24de52faac24584df