Updated openstack/openstack Project: openstack/cinder  966a1fdb2b403a0c521a26d8c5a262da24131354
Update generate_cinder_opts path

The location of the generate_cinder_opts script was moved in
d3abafdee4affe457e39705fd9b8e27b0f80f657. There is a check in
the script to exclude the script to its parsing, but this path
was not updated when it was moved. This updated the path to
the new location.

Change-Id: I369dec534c69f67fea72959bc9d626dddd8768ff