Updated openstack/openstack Project: openstack/neutron  8142957f53a6f7cfee4c7b9af22dfb43828323ed
Don't trigger DVR port update if status the same

The DVR path was missing logic to avoid triggering DB updates
even if the status was already correct. This was leading to a
constant back-and-forth updating between the agent and the server
effectively not changing anything but the updated_at timestamp
and the revision_number.

Closes-Bug: #1716522
Change-Id: I88c678ac8a562db91c71edbc71dc43d1ada6c452