Updated openstack/openstack Project: openstack/tempest  2c8e377d6338d3bccb1cc9e540fa344d8e00c0a1
Add a validation resources fixture

Make validation resources available via a fixture that can be used
by tests conveniently via the useFixture helper.
Add unit tests for it and the validation resources module in general.

This work in preparation to move validation resources to lib.

This commit introduces the RegistryFixture which sets up a usable
registry to unit tests of features that do not rely on configuration
but need a registry. The fixture cleans up the registry once the
test is over and this uncovered the fact that a few tests
erroneously dependent on the registry being populated - which just
happened to be because of some other tests.

Which the new registry fixture in use, such unit tests have to be
fixed in this patch.

Change-Id: I0faaff07e446e35537b53289fc8980f6692fcf55