Updated openstack/openstack Project: openstack/swift  7caed1a9df69f0ee2dced0c3365a03943c2c093b
Make swift-dispersion-report importable

This patch allows to import the dispersion report tool, and thus making
it more easily usable within other Python tools. This can be also used
in a follow up patch to add some tests for the report tool.

It also fixes a bug when using the "--dump-json" option - until now it
returned the policy name and made the JSON invalid.

Change-Id: Ie0d52a1a54fc152bb72cbb3f84dcc36a8dad972a