Updated openstack/openstack Project: openstack-infra/project-config  19247d92c39e8272fae258ea24d945219a4edcd9
Make wheel mirror builders normal zuul nodes

Including puppet madness!

We don't need the long-lived wheel builders - we build from scratch
every time. So all we need is AFS running on a node.

AFS, however, is tricky on CentOS. Rewriting all that right now is meh.
We can do that later.

This gets us to not needing special nodes for wheel builders. Yay!

Change-Id: I6d92c47f4f7a26ca290cd400452cb0a73872b24d
NOTE: It may be inefficient to run this on our low-bandwidth nodes.
Signed-off-by: Paul Belanger <pabelanger@redhat.com>