Updated openstack/openstack Project: openstack/barbican  1d20294b994af9abfcc4292bb79fcd32581d0c32
Add extra time in functional test that fails intermittently

The functional test that checks secret expiration fails
intermittently because sometimes, when the gate is slow,
the secret expires too quickly for the test.

This patch adds an extra 10 seconds so the test will pass
more consistently, while still maintaining the integrity
of the test.

Change-Id: I2f0df9b42dd2bf9dd600948164532fd31bb2a0d3
Closes-Bug: #1499673