Updated openstack/openstack Project: openstack/neutron-fwaas  7ddc450eebed744d01d1e370ee32dfda8e9a54ea
Fix mismatch in error messages

Commit [1] changed an exception message and that will lead to a unit test
failure when a new version of neutron-lib is released. Strictly speaking
the change introduced is non backward compatible, but the nature of the
issue (typo) does not justify any effort except removing a brittle assert
in the unit test. Validating the conflict error is more than enough.

[1] 26de187f6036481b20a2d1c22a13fbc4db5c954b

Change-Id: I3b8516582b95cc10b1f691c7538cb7382a8f8a07