Updated openstack/openstack Project: openstack/neutron  ccfa93a585846b847117feb4e1e05aa788d591dd
Switch to Pecan for unit tests

This will drop Pecan in place of the old APIRouter for all
unit tests. This will significantly increase the UT coverage
of pecan and will allow us to drop the old API routing logic
completely.

The rest of the inline test changes are due to slight plugin loading
differences.

Implements: blueprint wsgi-pecan-switch
Change-Id: I76dc23fb7b96d82b0da50285bd0aac76142e81e5