Updated openstack/openstack Project: openstack/castellan  12f69f7c1cbcae8906a677a212515d9b37a37bce
Makes list method not abstract

Any implementations of key_manager that don't have "list"
defined (i.e. ConfKeyManager in Nova and Cinder) will not be
instantiable if they try to use a version of Castellan
that was released after "list" was added. Adds a default
implementation of "list" that returns nothing for backwards
compatibility.

Closes-Bug: #1715451
Change-Id: I1e413831163bffaed3a2580f039e242da7d303f8