Updated openstack/openstack Project: openstack/nova  91addc87c67c83b59bbe336e8ecfb80f784cdfbc
Remove deprecated keymgr code

Now that enough time has passed, the keymgr code that was
deprecated for removal can be removed.

Barbican is the default option for Castellan, but Barbican is not
part of default DevStack yet. Until Barbican is used by default in
the dsvm gates, ConfKeyManager (the fixed_key key manager) should
be set in DevStack, which was added with
I733279864ee1a4aaffc9c8eed81b5e12f8d8821b.

Change-Id: I82ee74f3d2629281dc8116af55f6a7b5398fc473