Updated openstack/openstack Project: openstack/tempest  2c8db1a8053b694756034fd250d8cf1611937232
Prepare network_resources as a stable interface

The network_resources attribute should not be a public interface,
so renaming it to _network_resources. Document set_network_resources
better; ensure it's only invoked before super's setup_credentials is
invoked to avoid the call being ignored and generate errors that
may be difficult to debug.

Change-Id: I4eab8f2a722b47edc20e4aab0ef453bec16842f3