Updated openstack/openstack Project: openstack/tempest  df7c46fdbfb0f00be8c4f5d38d04003c36834116
Make skip_checks and setup_creds safe

The skip_checks and setup_credentials "fixtures" are implemented
in the base class. The implementation can be extended by test
classes but it should not be overwritten.

Enforcing this by checking that the base class version has been
invoked. Added unit tests for it as well.

Change-Id: Ie5d994519f9d56de423d1a0bede7ee8d703b9390