Updated openstack/openstack Project: openstack/nova  d1e70e980113ef8b7aacb63a31345a9275e237cf
Default the service version in the notification tests

The service.update notification contains the version of the service.
This causes that at every service version update the service.update
notification sample needs to be updated as well.

This patch changes the sample matching in the service notification
tests to default the expected service version to the current service
version. This way the sample doesn't need to be update in the future
due to service version change.

Change-Id: I7f2485c01ae44f6061cba86f88ac9614a1dc3295