Updated openstack/openstack Project: openstack/requirements  ee1a3433f13f405cfb7871573c4e1ebbf31f98ac
Xfail the mitmproxy / pyOpenSSL dependency conflict

This is already fixed in master:
https://github.com/mitmproxy/mitmproxy/pull/2435

So it'll hopefully in the next release.

Because the pyOpenSSL package name and requirement name do not match we
also add code to lower the name for comparison.

Also the check-uc only complains about a single issue but there are 3.

Change-Id: I6657802908ac4052eeb6c73fb03098b0c3557eb2