Updated openstack/openstack Project: openstack/nova  4932720dad43d7e20949c575b1fea61750937468
Target context when setting instance to ERROR when over quota

When conductor does the quota recheck, the instances are created
in a cell but when we update the instance and set it to ERROR state,
we were not targeting the context to the cell that the instance lives
in, which leads to an InstanceNotFound error and then the instance
is stuck in BUILD/scheduling state.

This targets the context to the cell when updating the instance.

Change-Id: I45faffaba4d329433a33cfb5e64c89ce4885df46
Closes-Bug: #1717000