Updated openstack/openstack Project: openstack/cinder  6538682ecb19052d80519a0d9e84627cd97d1dea
Use conditional update for group update and delete

Consistency groups had conditional updating to handle API race
conditions, but the switch to groups did not include that.

Adding conditional update handling for update and delete so we
have the same protection. Also relaxed the restriction on update
to allow updating name or description when in states other than
Available.

Change-Id: I9ddd7e881be23be8b7d37063d87417c47deda9e8
Closes-bug: #1673319