Updated openstack/openstack Project: openstack/oslo.messaging  393ecff3451091404832dd6b8a088e1bec760101
Warn when wrong transport instance is used

Since RPC and notifications can have different backends, it is
useful to warn users if they use a notification transport in RPC
and vice versa. This patch introduces RPCTransport and
NotificationTransport subclasses of Transport, so it's easier to
add different behavior for them if need be.

Related-Bug: #1680192
Change-Id: Iab60544d69053c8e74c28a2d5c84665be749013f