Updated openstack/openstack Project: openstack/horizon  177754b8ef4e5cce0d8d8c8d9da199efe45b6dd3
Fix link to serial console stylesheet

The serial_console.css stylesheet was renamed to serial_console.scss in
commit 09706c6e3754c218b00dcaf210629aa6970dc80a, but the link
referencing it in the serial console template was not updated. As a
result the serial console is displayed with a proportional font.

This commit updates the stylesheet link and adds a compress block to
compile it from SCSS to CSS.

Change-Id: Ic8abc970e76cb67c79c2335370c63e235a23eb5a
Closes-Bug: #1714311