Updated openstack/openstack Project: openstack-infra/project-config  4945efa54d1a420843edec122038cc0840022a0e
Add 10 minutes to linuxbridge job

Looking at a recent timeout at [1], the linuxbridge job was a few
tests away from completing when the job timed out. The root cause
of the timeout was actually devstack-gate-setup-workspace-new taking
22 minutes instead of the 3-5 it should have which ate up a significant
amount of time.

I analyzed the difference between the execution time of the actual test
suite run and the a test suite run on the OVS job and it took about 5
minutes longer.

Adding 10 minutes should be enough to compensate for landing on a slow
node with a long setup time and should bring timeouts to parity with the
OVS job.

1. http://logs.openstack.org/75/502975/1/gate/
   gate-tempest-dsvm-neutron-linuxbridge-ubuntu-xenial/13d6c14/

Change-Id: Ieededb459e884e2a158678da7b024549423a76b1
Related-Bug: #1660612