Updated openstack/openstack Project: openstack-infra/project-config  bb3173e2271f73db7c5d1ebaa271e8f515db6704
Add validate-host to base-minimal

The intent of leaving validate-host out of base minimal was to
run integration tests on it.
However, validate-host currently contains tasks that can't be
tested outside of a trusted context due to tasks that require
running on the executor.

validate-host is the role that creates the zuul-info directory
with the job inventory and variables, making it valuable to
troubleshoot integration jobs. Let's add it.

Change-Id: Ifad37daf1db116e29cbf18b242dff2959bce0331