Updated openstack/openstack Project: openstack/castellan  44ea22891d6707fcf3c55517a6adf33c13e2a913
allow redirects in .htaccess files on the static web servers

As part of allowing projects to establish redirects when they move
documentation pages around, we need to let them use .htaccess files.
We only want to allow Redirect and RedirectMatch directives,
though. This change restricts the directives on docs.o.o and adds them
on the static servers used for drafts published during a build (to
allow testing of the redirects).

For more details see:
 http://lists.openstack.org/pipermail/openstack-dev/2017-July/120240.html

Change-Id: I256d69f08499549e9e21179eb89f49cda79f7ac7