Update generate-gitmodules script for official deliverables
We should generate the list of deliverables from the projects.yaml file
in the governance repo. We should also make the submodule entries
relative rather than absolute so that browsing in other places works.

Also, when adding a git submodule, one needs to use the git submodule
add command. So do that, and trim out old no-longer-existing submodules
too.

Change-Id: I97c812f8497cc32a2ee1b418fce2e6bd20af3bc3